@@ -1,5 +1,19 @@
-"""MCP Gateway - Universal bridge for Claude Desktop"""
+"""MCP Gateway - Universal bridge for Claude Desktop.
+
+A Python 3.11+ gateway providing seamless integration between Claude Desktop
+and various system automation tools via the Model Context Protocol (MCP).
+"""
+
+from __future__ import annotations
 
 __version__ = "1.0.0"
-__author__ = "MCP Gateway Team"
+__author__ = "Robert (Masa) Matsuoka"
+__email__ = "bobmatnyc@users.noreply.github.com"
 __license__ = "MIT"
+
+__all__ = [
+    "__version__",
+    "__author__",
+    "__email__",
+    "__license__",
+]
